+++ README.md
@@ -138,7 +138,7 @@ function getAllAddons({ categoryId }: GetAllAddonsParams = {}) {
   'Inexact type is incompatible with exact type' but that's a
   [bug](https://github.com/facebook/flow/issues/2405).
   You can use the `Exact<T>` workaround from
-  [`src/core/types/coreTypes`](https://github.com/mozilla/addons-frontend/blob/master/src/core/types/coreTypes.js)
+  [`src/core/types/util`](https://github.com/mozilla/addons-frontend/blob/master/src/core/types/util.js)
   if you have to. This is meant as a working replacement for
   [$Exact<T>](https://flow.org/en/docs/types/utilities/#toc-exact).
 * Try to avoid loose types like `Object` or `any` but feel free to use
