+++ src/core/types/util.js
